requirements.txt
fastapi==0.115.6
SQLAlchemy==2.0.36
alembic==1.14.0
pydantic[email]==2.10.4
bcrypt==4.2.1
passlib==1.7.4
python-jose==3.3.0
websockets==14.1
aiosqlite==0.20.0
Jinja2==3.1.5
simple_py_config==0.0.1
# (.venv) dima@dima-Nitro-AN515-52:~/python/projects/fastapi-mini-chat$ pip freeze
# aiosqlite==0.20.0
# alembic==1.14.0
# annotated-types==0.7.0
# anyio==4.7.0
# bcrypt==4.2.1
# dnspython==2.7.0
# ecdsa==0.19.0
# email_validator==2.2.0
# fastapi==0.115.6
# greenlet==3.1.1
# idna==3.10
# Jinja2==3.1.5
# Mako==1.3.8
# MarkupSafe==3.0.2
# passlib==1.7.4
# pyasn1==0.6.1
# pydantic==2.10.4
# pydantic_core==2.27.2
# python-jose==3.3.0
# rsa==4.9
# simple_py_config==0.0.1
# six==1.17.0
# sniffio==1.3.1
# SQLAlchemy==2.0.36
# starlette==0.41.3
# typing_extensions==4.12.2
# websockets==14.1